--- a/src/GraphQL.Tests/Types/ObjectGraphTypeTests.cs
+++ b/src/GraphQL.Tests/Types/ObjectGraphTypeTests.cs
@@ -34,7 +34,7 @@ namespace GraphQL.Tests.Types
         [Fact]
         public void should_throw_on_invalid_graphtype_name()
         {
-            var ex = new ArgumentOutOfRangeException("name", "A type name must match /^[_a-zA-Z][_a-zA-Z0-9]*$/ but ::: does not.");
+            var ex = new ArgumentOutOfRangeException("name", "A type name must match /^[_a-zA-Z][_a-zA-Z0-9]*$/ but ':::' does not.");
             Should.Throw<ArgumentOutOfRangeException>(() => new TypeWithInvalidName())
                 .Message.ShouldBe(ex.Message);
         }