--- a/src/GraphQL.Tests/Utilities/NameValidatorTests.cs
+++ b/src/GraphQL.Tests/Utilities/NameValidatorTests.cs
@@ -1,7 +1,7 @@
 using System;
+using GraphQL.Utilities;
 using Shouldly;
 using Xunit;
-using static GraphQL.Utilities.NameValidator;
 
 namespace GraphQL.Tests.Utilities
 {
@@ -9,29 +9,36 @@ namespace GraphQL.Tests.Utilities
     {
         [Fact]
         public void ValidateName_whenNameIsNull_throwsArgumentOutOfRange() =>
-           Should.Throw<ArgumentOutOfRangeException>(() => ValidateName(null));
+           Should.Throw<ArgumentOutOfRangeException>(() => NameValidator.ValidateName(null));
 
         [Fact]
         public void ValidateName_whenNameIsEmpty_throwsArgumentOutOfRange() =>
-            Should.Throw<ArgumentOutOfRangeException>(() => ValidateName(string.Empty));
+            Should.Throw<ArgumentOutOfRangeException>(() => NameValidator.ValidateName(string.Empty));
 
         [Fact]
         public void ValidateName_whenNameIsWhitespace_throwsArgumentOutOfRange() =>
-            Should.Throw<ArgumentOutOfRangeException>(() => ValidateName(" "));
+            Should.Throw<ArgumentOutOfRangeException>(() => NameValidator.ValidateName(" "));
 
         [Fact]
         public void ValidateName_whenNameStartsWithReservedCharacters_throwsArgumentOutOfRange() =>
-            Should.Throw<ArgumentOutOfRangeException>(() => ValidateName("__dede"));
+            Should.Throw<ArgumentOutOfRangeException>(() => NameValidator.ValidateName("__dede"));
 
-        [Fact]
-        public void ValidateName_whenNameContainsInvalidCharacters_throwsArgumentOutOfRange() =>
-            Should.Throw<ArgumentOutOfRangeException>(() => ValidateName("śćłó"));
+        [Theory]
+        [InlineData("śćłó")]
+        [InlineData("3test")]
+        [InlineData("test Name")]
+        public void ValidateName_whenNameContainsInvalidCharacters_throwsArgumentOutOfRange(string invalidName) =>
+            Should.Throw<ArgumentOutOfRangeException>(() => NameValidator.ValidateName(invalidName));
 
-        [Fact]
-        public void ValidateName_whenNameIsCorrect_DoesntthrowsArgumentOutOfRange()
+        [Theory]
+        [InlineData("goodName")]
+        [InlineData("name3")]
+        [InlineData("Test")]
+        [InlineData("Test_Name")]
+        [InlineData("_test")]
+        public void ValidateName_whenNameIsCorrect_DoesntthrowsArgumentOutOfRange(string validName)
         {
-            var ex = Record.Exception(() => ValidateName("goodName"));
-            ex.ShouldBeNull();
+            NameValidator.ValidateName(validName);
         }
     }
 }