--- a/src/GraphQL/Utilities/StringUtils.cs
+++ b/src/GraphQL/Utilities/StringUtils.cs
@@ -1,72 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text.RegularExpressions;
 
 namespace GraphQL.Utilities
 {
     public static class StringUtils
     {
-        private static readonly Regex _words = new Regex(@"[A-Z\xc0-\xd6\xd8-\xde]?[a-z\xdf-\xf6\xf8-\xff]+(?:['’](?:d|ll|m|re|s|t|ve))?(?=[\xac\xb1\xd7\xf7\x00-\x2f\x3a-\x40\x5b-\x60\x7b-\xbf\u2000-\u206f \t\x0b\f\xa0\ufeff\n\r\u2028\u2029\u1680\u180e\u2000\u2001\u2002\u2003\u2004\u2005\u2006\u2007\u2008\u2009\u200a\u202f\u205f\u3000]|[A-Z\xc0-\xd6\xd8-\xde]|$)|(?:[A-Z\xc0-\xd6\xd8-\xde]|[^\ud800-\udfff\xac\xb1\xd7\xf7\x00-\x2f\x3a-\x40\x5b-\x60\x7b-\xbf\u2000-\u206f \t\x0b\f\xa0\ufeff\n\r\u2028\u2029\u1680\u180e\u2000\u2001\u2002\u2003\u2004\u2005\u2006\u2007\u2008\u2009\u200a\u202f\u205f\u3000\d+\u2700-\u27bfa-z\xdf-\xf6\xf8-\xffA-Z\xc0-\xd6\xd8-\xde])+(?:['’](?:D|LL|M|RE|S|T|VE))?(?=[\xac\xb1\xd7\xf7\x00-\x2f\x3a-\x40\x5b-\x60\x7b-\xbf\u2000-\u206f \t\x0b\f\xa0\ufeff\n\r\u2028\u2029\u1680\u180e\u2000\u2001\u2002\u2003\u2004\u2005\u2006\u2007\u2008\u2009\u200a\u202f\u205f\u3000]|[A-Z\xc0-\xd6\xd8-\xde](?:[a-z\xdf-\xf6\xf8-\xff]|[^\ud800-\udfff\xac\xb1\xd7\xf7\x00-\x2f\x3a-\x40\x5b-\x60\x7b-\xbf\u2000-\u206f \t\x0b\f\xa0\ufeff\n\r\u2028\u2029\u1680\u180e\u2000\u2001\u2002\u2003\u2004\u2005\u2006\u2007\u2008\u2009\u200a\u202f\u205f\u3000\d+\u2700-\u27bfa-z\xdf-\xf6\xf8-\xffA-Z\xc0-\xd6\xd8-\xde])|$)|[A-Z\xc0-\xd6\xd8-\xde]?(?:[a-z\xdf-\xf6\xf8-\xff]|[^\ud800-\udfff\xac\xb1\xd7\xf7\x00-\x2f\x3a-\x40\x5b-\x60\x7b-\xbf\u2000-\u206f \t\x0b\f\xa0\ufeff\n\r\u2028\u2029\u1680\u180e\u2000\u2001\u2002\u2003\u2004\u2005\u2006\u2007\u2008\u2009\u200a\u202f\u205f\u3000\d+\u2700-\u27bfa-z\xdf-\xf6\xf8-\xffA-Z\xc0-\xd6\xd8-\xde])+(?:['’](?:d|ll|m|re|s|t|ve))?|[A-Z\xc0-\xd6\xd8-\xde]+(?:['’](?:D|LL|M|RE|S|T|VE))?|\d+|(?:[\u2700-\u27bf]|(?:\ud83c[\udde6-\uddff]){2}|[\ud800-\udbff][\udc00-\udfff])[\ufe0e\ufe0f]?(?:[\u0300-\u036f\ufe20-\ufe23\u20d0-\u20f0]|\ud83c[\udffb-\udfff])?(?:\u200d(?:[^\ud800-\udfff]|(?:\ud83c[\udde6-\uddff]){2}|[\ud800-\udbff][\udc00-\udfff])[\ufe0e\ufe0f]?(?:[\u0300-\u036f\ufe20-\ufe23\u20d0-\u20f0]|\ud83c[\udffb-\udfff])?)*");
-
-        /// <summary>
-        /// Split a cased string into a series of "words" excluding the separator if applicable.
-        /// </summary>
-        public static IEnumerable<string> ToWords(string str)
-        {
-            foreach (Match match in _words.Matches(str))
-            {
-                yield return match.Value;
-            }
-        }
-
-        public static string Capitalize(string str)
-        {
-            switch (str.Length)
-            {
-                case 0:
-                    return str;
-                case 1:
-                    return str.ToUpperInvariant();
-                default:
-                    str = str.ToLowerInvariant();
-                    return char.ToUpperInvariant(str[0]) + str.Substring(1);
-            }
-        }
-
-        public static string ToCamelCase(string str)
-        {
-            return ChangeCase(str, "", (word, index) => index == 0 ? word.ToLowerInvariant() : Capitalize(word));
-        }
-
-        public static string ToConstantCase(string str)
-        {
-            return ChangeCase(str, "_", w => w.ToUpperInvariant());
-        }
-
-        public static string ToPascalCase(string str)
-        {
-            return ChangeCase(str, "", Capitalize);
-        }
-
-        public static string ChangeCase(string str, string sep, Func<string, string> composer)
-        {
-            return ChangeCase(str, sep, (w, i) => composer(w));
-        }
-
-        public static string ChangeCase(string str, string sep, Func<string, int, string> composer)
-        {
-            string result = "";
-            int index = 0;
-
-            foreach (string word in ToWords(str))
-            {
-                result += (index == 0 ? "" : sep) + composer(word, index++);
-            }
-
-            return result;
-        }
-
         /// <summary>
         /// Given [ A, B, C ] return '"A", "B", or "C"'.
         /// </summary>
@@ -107,13 +46,6 @@ namespace GraphQL.Utilities
             });
 
             return optionsByDistance.OrderBy(x => x.Value).Select(x => x.Key).ToArray();
-        }
-
-        public static void Swap<T>(ref T arg1, ref T arg2)
-        {
-            T temp = arg1;
-            arg1 = arg2;
-            arg2 = temp;
         }
 
         /// <summary>
@@ -197,6 +129,13 @@ namespace GraphQL.Utilities
 
             int result = dCurrent[maxi];
             return result > threshold ? int.MaxValue : result;
+
+            static void Swap<T>(ref T arg1, ref T arg2)
+            {
+                T temp = arg1;
+                arg1 = arg2;
+                arg2 = temp;
+            }
         }
     }
 }