I ran some performance tests. It only really helps people with transient/scoped schemas...but it does help.  What do you think?

```
Before: (transient schema)

|        Method |     Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |---------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection | 5.842 ms | 0.0275 ms | 0.0257 ms | 484.3750 | 31.2500 |     - |    2.9 MB |
|          Hero | 4.563 ms | 0.0329 ms | 0.0292 ms | 367.1875 | 31.2500 |     - |    2.2 MB |

After: (transient schema)

|        Method |     Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |---------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection | 5.751 ms | 0.0460 ms | 0.0408 ms | 476.5625 | 31.2500 |     - |   2.87 MB |
|          Hero | 4.409 ms | 0.0335 ms | 0.0297 ms | 359.3750 | 31.2500 |     - |   2.17 MB |
```