@Shane32 2021-01-06
FYI, if you compare the ToConstantCase method against the regex equivalent,

```

|          Method |       Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |-----------:|----------:|----------:|-------:|------:|------:|----------:|
|           RegEx | 5,721.6 ns | 110.25 ns | 122.55 ns | 0.3662 |     - |     - |    1592 B |
| CustomFunction1 |   408.7 ns |   4.79 ns |   4.48 ns | 0.2027 |     - |     - |     848 B |
| CustomFunction2 |   267.9 ns |   1.79 ns |   1.59 ns | 0.2027 |     - |     - |     848 B |
```

`CustomFunction2` looks for A-Z and a-z, etc, whereas `CustomFunction1` checks `char.IsUpper` and `char.IsLower`