@Shane32 2021-01-06
By the way, even though this implementation is 20x faster, it could be a lot faster yet, even without using System.Buffers, which could make it faster still.