@sungam3r 2021-01-06
> What do you think?

I like the changes. What is your confidence that you can just delete all the calls to `TrimGraphQLTypes`? I wanted to do it myself some time ago in one of the PRs, when I was discussing the possibility of finding special characters in the name. Apparently they shouldn't be there, but I didn't dare to just delete the calls. 

For the rest of the edits, I can say thank you for removing this garbage.