@Shane32 2021-01-07
> I like the changes. What is your confidence that you can just delete all the calls to `TrimGraphQLTypes`? I wanted to do it myself some time ago in one of the PRs, when I was discussing the possibility of finding special characters in the name. Apparently they shouldn't be there, but I didn't dare to just delete the calls.

I'm very confident.  However, just in case there is some issue we are unaware of, I think we should put this in 4.0.  I don't see any reason to work on 3.3 when there may be a question of how it functions, and especially since we are so close to releasing 4.0.  We want the 3.2 -> 3.3 transition to be 100% seamless.

If someone is going to go through the effort of migrating to 4.0, and they encounter a bug, so be it.  We can fix it and will just release a 4.0.1.  No big deal.  Besides, we can test 4.0 beta ourselves with our own large scale projects if we want extra confidence that everything is good.  But I'm pretty confident that the call can be removed entirely.

Besides all that, with a singleton schema, it doesn't matter how long it takes the the function to execute.  It's just for the few people that have scoped schemas.  So there's no reason to put these optimizations into 3.3