@Shane32 2021-01-07
I'll postpone this after we merge changes for ToCamelCase.  Or, if you want to pick up this PR that's fine too.  I was just trying to get rid of regex, knowing how slow it is.