@Shane32 2021-01-08
@sungam3r If you want to finish this PR up, rewriting it with stackalloc or whatever, let me know. I've just been a bit busy.