@Shane32 2021-01-08
> Or we leave this algorithm as it is (enriching with comments), given that its results are still cached.

Done.  Simplified the algorithm a little so it's easier to follow also.